+++ clang-tools-extra/clang-tidy/daedalean/OperatorOverloadingCheck.h
@@ -1,0 +1,38 @@
+//===--- OperatorOverloadingCheck.h - clang-tidy ----------------*- C++ -*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_DAEDALEAN_OPERATOROVERLOADINGCHECK_H
+#define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_DAEDALEAN_OPERATOROVERLOADINGCHECK_H
+
+#include "../ClangTidyCheck.h"
+
+namespace clang {
+namespace tidy {
+namespace daedalean {
+
+/// Daedalean CS.R.41 Operator overloading
+///
+/// For the user-facing documentation see:
+/// http://clang.llvm.org/extra/clang-tidy/checks/daedalean-operator-overloading.html
+class OperatorOverloadingCheck : public ClangTidyCheck {
+public:
+  OperatorOverloadingCheck(StringRef Name, ClangTidyContext *Context)
+      : ClangTidyCheck(Name, Context) {}
+  void registerMatchers(ast_matchers::MatchFinder *Finder) override;
+  void check(const ast_matchers::MatchFinder::MatchResult &Result) override;
+
+private:
+  void checkOperatorPair(const FunctionDecl * function, OverloadedOperatorKind kind, const std::string & message);
+  void checkReturnSelf(const Stmt *stmt);
+};
+
+} // namespace daedalean
+} // namespace tidy
+} // namespace clang
+
+#endif // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_DAEDALEAN_OPERATOROVERLOADINGCHECK_H